Bug fix:
When generating task topics, if the file has multiple ordered lists, but
the list are not separated by headers of any kind, the tool scrambles
the contents of the topic.

This commit fixes this with an additional check: If a topic contains > 1
steps element, it preserves the first one and converts everything else
to a regular ordered list.